app/requirements.txt
selenium
pylint
python-dotenv